migration solve

+++ database/migrations/2024_09_05_075239_create_employees_table.php
@@ -13,9 +13,9 @@ return new class extends Migration
     {
         Schema::create('employees', function (Blueprint $table) {
             $table->id();
-            $table->integer('nip');
+            $table->bigInteger('nip');
             $table->string('name');
-            $table->integer('no_handphone');
+            $table->bigInteger('no_handphone');
             $table->foreignId('user_id')->constrained('users')->cascadeOnDelete();
             $table->foreignId('jabatan_id')->constrained('jabatans')->cascadeOnDelete();
             $table->foreignId('team_id')->constrained('teams')->cascadeOnDelete();
